+++ src/Public/Join-File.ps1
@@ -57,35 +57,36 @@ function Join-File {
                     New_ArgumentException 'The argument specified must resolve to a valid path on the FileSystem provider.' -Throw
                 }
 
+                [System.Collections.Generic.Dictionary[string, System.IDisposable]] $Disposable = @{}
+
                 [System.IO.FileInfo] $File = $Object.ProviderPath
 
-                [string] $CalculatedDestination = $DestinationInfo.Extension | Use-Ternary $DestinationInfo "$( $DestinationInfo.FullName.TrimEnd('\/') )/$( $File.BaseName )"
+                [string] $CalculatedDestination = if ($DestinationInfo.Extension) { $DestinationInfo } else { "$( $DestinationInfo.FullName.TrimEnd('\/') )/$( $File.BaseName )" }
 
                 if ($PSCmdlet.ShouldProcess($CalculatedDestination, 'Write Content')) {
-                    if (-not ([string] $Directory = $DestinationInfo.Extension | Use-Ternary $DestinationInfo.Directory $DestinationInfo).Exists) {
-                        $null = [System.IO.Directory]::CreateDirectory($Directory)
-                    }
+                    [string] $Directory = if ($DestinationInfo.Extension) { $DestinationInfo.Directory } else { $DestinationInfo }
+
+                    if (-not $Directory.Exists) { $null = [System.IO.Directory]::CreateDirectory($Directory) }
 
                     $PSCmdlet.WriteVerbose("WRITE ${CalculatedDestination}")
 
-                    Use-Object ([System.IO.FileStream] $Writer = [System.IO.File]::OpenWrite($CalculatedDestination)) {
-                        # sort to fix ChildItem number sorting
-                        foreach ($SplitFile in (Get-ChildItem -Path "$( $File.Directory )/$( $File.BaseName ).*split").FullName | Sort-Object -Property { [int32] [regex]::Match($_, '\.(?<match>\d+)split$').Groups['match'].Value }) {
-                            $PSCmdlet.WriteVerbose("READ ${SplitFile}")
+                    $Disposable.Writer = [System.IO.File]::OpenWrite($CalculatedDestination)
+                    # sort to fix ChildItem number sorting
+                    foreach ($SplitFile in (Get-ChildItem -Path "$( $File.Directory )/$( $File.BaseName ).*split").FullName | Sort-Object -Property { [int32] [regex]::Match($_, '\.(?<match>\d+)split$').Groups['match'].Value }) {
+                        $PSCmdlet.WriteVerbose("READ ${SplitFile}")
 
-                            [byte[]] $Bytes = [System.IO.File]::ReadAllBytes($SplitFile)
-                            $Writer.Write($Bytes, 0, $Bytes.Length)
-                        }
+                        [byte[]] $Bytes = [System.IO.File]::ReadAllBytes($SplitFile)
+                        $Disposable.Writer.Write($Bytes, 0, $Bytes.Length)
                     }
                 }
 
                 $PSCmdlet.WriteObject(( Get-ChildItem -Path $CalculatedDestination ))
-
-                ## EXCEPTIONS ##################################################
             } catch [System.Management.Automation.MethodInvocationException] {
                 $PSCmdlet.WriteError(( New_MethodInvocationException -Exception $_.Exception.InnerException ))
             } catch {
                 $PSCmdlet.WriteError($_)
+            } finally {
+                $Disposable.Values.Dispose()
             }
         }
     }