+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 - Fix Linux Compatibility and Test Suite [!35](https://gitlab.com/PoshAJ/PoshToolbox/-/merge_requests/35)
 - General Format and Consistency [!39](https://gitlab.com/PoshAJ/PoshToolbox/-/merge_requests/39)
+- Fix Function Execution Inside Module Scope [!41](https://gitlab.com/PoshAJ/PoshToolbox/-/merge_requests/41)
 
 ### Changed